create and test all api product + user

+++ controller/user.js
@@ -5,7 +5,9 @@ module.exports = {
         const result = await User.create({
             name: req.body.name,
             age: req.body.age,
-            phone: req.body.phone
+            phone: req.body.phone,
+            username: req.body.username,
+            password: req.body.password
         })
         return res.json(result);
     },
@@ -13,16 +15,16 @@ module.exports = {
         const result = await User.find()
         return res.json(result);
     },
-    deleteUser: async (req, res) => {
-        await User.deleteOne();
-    },
     findUser: async (req, res) => {
-        const result = await User.where();
-        query.findOne()
+        const result = await User.findOne()
         return res.json(result);
     },
     deleteUser: async (req, res) => {
-        const reult = await User.deleteOne()
+        const result = await User.deleteOne({_id: req.params.userId})
+        return res.json(result);
+    },
+    updateUser: async (req, res) => {
+        const result = await User.updateOne({_id: req.params.userId}, {name: req.body.name, age: req.body.age, phone: req.body.phone})
         return res.json(result);
     }
 }
